requirements2.txt
panda
pymongo
redis
python-dotenv
pymongo
tqdm
numpy